Ticket support
When slaves can't be allocated, the allocation will be put in
a queue, and a ticket will be handed out, which allows the client
to claim the allocation once satisfied.

When agents check in, and if there are any pending allocations
in the queue, the agent will be used to satisfy the alloction,
and if there isn't one that can be satisifed, it's put in the
'available' queue.

All this must be atomic.