Parameters are now stored as YAML
Since the job server will need to be able to extract metadata
such as parameters, they can't be defined as Python as we can not
run arbitrary scripts on that server.

Storing them as YAML is more secure, and is not that bad either
as it's more readable as well.